Be able to move cards up and down with the keyboard
Part of story #20960 : Have shortcuts to move cards

While focus is on a card, hitting shift+up and shift+down will move the
card up and down in the cell.

If the card is the first in cell, it will be moved to the bottom of the
cell. If it is the last one, it will be moved to the top.

Change-Id: I9d1f3528fe1990d1ed95a44270813fedcefe002d