Add artifact closure action enabled in database
This is part of story #17183 close artifacts from commit messages

A new column named allow_artifact_closure is added in the SQL table
responsible for GitLab repository integration.

It's default value is set to 0 (disabled by default). To test this
patch, this must be manally set to 1 in databse. When it's done, the
artifact closure action must be working as before in the project for the
given GitLab repository.

Change-Id: I2ba817f1edd511e1aff8fa8ac23d3caae085da2e