Update the user agent associated with a session when a change is detected
This is useful for long-lived session where the browser is susceptible
to be updated.

The easiest way to test the change is to update your entry in the
session DB table and to check that it is updated back to the expected
value when you browse.

Part of request #21338: Add statistics in the site administration about
browser usages based on the active sessions

Change-Id: Ie43a75fcf14ecb228bc37397a88b0db718999054