Each GitLab integration is unique
This is part of story #17183 close artifacts from commit messages

A forgeupgrade has been added to move all the integration data in one
table to have each integration unique (repository data + webhook +
project).

Each commit, tag and MR info must be duplicated for each integration of
the GitLab repository

Old tables are not removed, they will be purged (if necessary) once all is OK.

What is done ?
--------------

* REST route GET projects/:id/gitlab_repository uses the new SQL tables
* Creation of the integration uses the new SQL tables
* Deletion of the integration uses the new SQL tables
* Edit integration API Token
* Regenerate Webhook Secret
* New webhook endpoint deals with Commit events
* New webhook endpoint deals with Tag events
* New webhook endpoint deals with MR events
* Old webhook endpoint is able to do all the actions as previously done
* Notifications are sent if needed

What is new ?
-------------

The big change is about the webhook configuration regeneration. When a
regeneration is done for an integration, the will delete in all the case
the webhook configuration on GitLab side and other intergrations of this
same GitLab repository will need to be reconfigured.

A visual warning is added to show that these integrations must
regenerate their webhook configuration.

Change-Id: If7be2225db1435dc83fac11a840247c559af9dc4