feat: Display tasks with end < start
In order to prevent users searching for tasks that have end date lesser
than start date, we now expose them in the gantt chart.

Those tasks are displayed at the end of tasks (or at the end of their
sibling subtasks) with a warning sign, but without any bar at all. The
popover displays an error message to explain such display.

Dependency links between those tasks and another tasks are not displayed
as well.

Part of story #20979: Display children in Roadmap

Change-Id: If58fc2c66d654f6c7f810b7304fbe1a394dab66e