Install and activate plugins recursively
The new CLI command will install (in the Tuleap sense)
AND activate any provided plugin. Unlike previous
implementation, it protects against typo in plugin names
and resolves dependencies automatically.

While the code theorically works for both the UI and
CLI command, front end still relies on dependency
checker. It might make sense to align the behaviour
later on.

Some people will be happy in the IT teams :)

Note: I had to remove the dependency on `botmattermost`
for `create_test_env`. There should not be any dependency
between a plugin that sits in the main repo and an external
plugin. In the present case, a future commit will remove
the feature (notifications in create_test_env) entirely.

Part of request #21334 enalean/tuleap-aio should be easier to test

Change-Id: I102d25bff23ae71dba62ca6c78346cd88595fa39