request #21366: Keep reverse cross reference when migrating a legacy tracker with its original IDs
When a legacy tracker is migrated with its original artifact IDs, the
reverse cross references are now kept.

To avoid the creation of some duplicate (notably for the artifacts cross
referencing another artifact in the same tracker) we clean up all the
existing cross references targetting an artifact in the migrated tracker
to recreate them from scratch.

Change-Id: Ie779c7c0557f4f59a3dad5f95b18147459091ee4