feat: dashed border when dates are implied
When artifact will have timeframe semantic based on linked artifact,
start dates and end dates of a task will be implied, based on another
artifact timeframe. In that case, the border of the Gantt bar will be dashed
instead of solid.

Since the new timeframe is not available yet, to test this contribution
you will have to edit the file `TaskRepresentationBuilderForTracker.php`
and use the following statements at L75:

```
$are_dates_implied = true;
```

Part of story #21362: Base artifact timeframe on a linked artifact

Change-Id: I617d7ebc416a8044ff28846d9d217bc17b29314f