chore: remove usage of TLP sass variable in siteadmin core
In order to be able to build only one stylesheet for BurningParrot core
pages, we need to convert usage of TLP sass variables to CSS variables.

Since BurningParrot main stylesheet is quite big and contains thousands
of conversion to be done, this patch focus only on siteadmin part. While
it does not reduce the number of stylesheet to build, it is part of
subsequents contributions that will let us reach that goal.

You can assert that the change is effective with the following command:

```
./node_modules/.bin/sass src/themes/BurningParrot/css/includes/_siteadmin.scss
```
It should not complain about unknown variables.

You should still rely on regular build of assets (cd src && npm run
build) to do the functional tests though: no changes are expected in
site administration.

Note: We could have split siteadmin scss from the rest of burning parrot
to have a dedicated build (and decrease the size of assets). However I
prefer to focus on the CSS variables conversion instead of searching
usage of random CSS selectors to not break something if by any luck they
are used outside of siteadmin.

Part of request #21354: Build only one stylesheet for BurningParrot core pages

Change-Id: Ia9cc1f5867a0c804464fd3df6779a35d58fd38db