request #21367: Disallow http:// URLs when adding a GitLab repository
Usage of http:// URLs, on top of being insecure, are very likely to not
work. Tuleap instances are expected to be served under HTTPS and
browsers will block/force upgrade to HTTPS the first request that is
done to retrieve the list of projects because it is active mixed
content.

If the GL server is not accessible with HTTPS users will get an error
which is more confusing than just preventing them to do the query (and
it is easier for the support team to troubleshoot).

Change-Id: Iec724369c61d5be222d6ea9d9c32570d76605785