Activate all plugins in enalean/tuleap-aio image
It's probably safe to assume that if RPMs were installed in the
image, it's because we want them to be installed at Tuleap side
and activated by default.

It won't be part of this contribution but it's likely that we
should actually add this as part of the default setup.

In order to test without activating too much crap on your server
you can test with tuleap-aio image itself:

    docker run --name tuleap-aio -v $PWD:/usr/share/tuleap enalean/tuleap-aio:centos7

Part of request #21334 enalean/tuleap-aio should be easier to test

Change-Id: I557578b09b7722e08def3c492b1deb4be87bc37c