request #21377: Optimize CompatPDODataAccess::escapeInt
When processing a large number of values this small change can save a
non negligible amount of time.

Ultimately this method should not be used at all and we should move
everything to prepared statements. Until we can reach this point, saving
a few CPU cycles on almost every requests is still nice to have 🌱.

The signature of ::escapeInt() has been updated to have the true return
value which has required a few fixes in the tracker and cardwall
plugins.

Change-Id: Ib472f58a3807a9ada4f542ba81699ee6bc1ee9fb