feat: Display accessibility patterns in Gantt bars
When user has activated the accessibility mode in user preferences,
patterns are displayed in the progress bar of tasks and subtasks so that
tasks from different trackers (with different colors) can be more easily
distinguished.

Please note that the patterns are not displayed for milestones and tasks
pwith 0% progress yet. The rationale is that we predict there won't be
two much of usage of both cases on one Gantt chart, it does not worth to
spend too much time on it.

Part of story #20979: Display children in Roadmap

Change-Id: I29ca90867e094109af3e84568aa86b2b0927189d