Add an ADR to explain mirror milestones
part of story #21368 create mirrored iteration

No functional change.

Notes:
In the epic, we explain _what_ mirror milestones are, _how_ they
function but never _why_ they were needed. It is a foundational
information to have in Program Management because a significant part of
the plugin's backend code is designed to handle them or modify them.

This ADR explains why Mirror Milestones are needed and the other
solution that was spiked at the beginning of "Multi-project backlog"
epic and why it was not chosen.

Change-Id: I7a17d3152bfc511ce701357e80d71ff498c85e1b