refactor: Move error/empty state in root state
In future contributions, we will fetch iterations to display them in
ribbons on top of tasks. Since this will share the error display, there
is no reason to keep it at the task state level. After all it is an
application state, it makes sense to put it in the root state.

For consistency, empty state is also moved in the root state.

No functional changes for display of Roadmap in error/empty states.

Part of story #21379: Display iterations on top of roadmap

Change-Id: I1ddc97de3f9e9da50c2d54b76c6b4fbe1f3677a3