request #21409 - TTM: Edit test button interactions should be consistent
In TestManagement, the `Edit test` button opens a modal to edit the
test.
However it is inside a `a` tag which leads to the artifact page of this
test.
Depending on where exactly the user clicks on the button, it will
trigger one or the other, which can be confusing.

This patch makes it so that left-clicking will always open the modal,
and middle-clicking will open the artifact page in a new tab.

It also adds `type="button"` to the test header buttons that don't submit
a form and `aria-hidden="true"` to icons that should not be taken
into account by assistive technologies.

Change-Id: I163f68d0c3e1ca76ca6161badcbcf5aee9818959