Transform Scrollbar/NavBar to TS + bump smooth-scrollbar
This is part of request #19729 Transform BurningParrot/ in core to TypeScript

The bump of smooth-scrollbar is because 7.x doesn't support TS, but
8.x yes.
So instead of using a module named `smooth-scrollbar.d.ts` to use the
lib in TS file, I decided to bump it.
You can find migration doc here [0].

What changes for us:
 => CCSS of smooth-scrollbar is now provided by JS.

No functional expected changes.
When you scroll, the navbar is still pinned (in tracker for example).

[0] https://github.com/idiotWu/smooth-scrollbar/blob/develop/docs/migration.md

Change-Id: I8db7e31ad02b59d5333c6acc81bed933048f8221