Run the PHP unit tests with PHP 8.0 in the nightly pipeline
You can also test the execution locally with `make tests-unit-php PHP_VERSION=80`.

Part of request #21412: Run PHP unit tests on PHP 8.0 nightly

This contribution does not close the request because while it works,
it's only because the build phase is done with PHP 7.4 so Composer does
not complain. We should still bump our dependencies before closing the
request.

Change-Id: I17582ee3893c98e30c65b1bd8fdfa0a9f5a39229