refactor: Introduce loadRoadmap
In the upcoming contribution, the display of the app will not only load
the tasks of the roadmap, but also its iterations. Therefore we need a
loadRoadmap that will be in charge of loading everything directly
related to the roadmap.

No functional change.

Part of story #21379: Display iterations on top of roadmap

Change-Id: Ifb08d4a24746f1996d39fdd8e96c33f26ca06c4c