feat: Parallel retrieval of tasks and iterations
If a roadmap has defined one or two iteration tracker (you have to hack
the db for now), then tasks and iterations are loaded in parallel.

* User see a loading state until all requests are completed
* User see an error state as soon as one request is in error

Currently, tasks and iterations have the same importance for decision
making, if one of the data is not available we prefer to display an
error state than an half built gantt chart. Furthermore if tasks are
displayed and suddenly an error state is displayed because iterations
are in error, it would be weird. So for now, it is all or nothing.

No visible change in the display of the widget for everybody.

Part of story #21379: Display iterations on top of roadmap

Change-Id: Idea9f01d0da24af38101ea982aa4dfba6c320452