Cannot create 'Scrum' or 'Agile ALM' projects
This issue was introduced by git #tuleap/stable/464e9fcce23cf35b3986ffd35ba31f81f3ff2439
and was caused by the fact the planning configuration now verifies if
the trackers used exist and is accessible by the user doing the change.

This verification causes an issue when the planning configuration is
done at the project creation since the project might be Pending and as
such not yet accessible to the user requesting the creation.

In order to solve the issue when we are dealing with a XML import of the
plannings we considerer the current user as a super user to bypass all the
restrictions. The only verification that is important when doing an XML
import is if the user has the right to do such import and this
verification is expected to be done once before starting the import.
It is also a good thing from a performance point of view since we do
less checks (and so less SQL queries) when importing the planning
configuration.

Closes request #21416: User can create invalid planning configurations

Change-Id: I316986092e54f154b4a7b43b6759b5743e95f4a9