feat: Display iterations ribbons
If a roadmap has defined one or two iteration trackers (you have to hack
the db for now), then corresponding ribbons are displayed.

To not clutter too much the review, for now it only displays dummy
"level 1"/"level 2", but later it will be replaced by real iterations.

There is no change for the display of subtask error messages, links,
popovers, …

Part of story #21379: Display iterations on top of roadmap

Change-Id: I202fbc6dc752ea46d5579d38d9680c378b8f0ba9