[feature] Artifact closure must only be done in default branch
This is part of story #21418 close artifacts with Done semantic

When a commit that asks for artifact closure is done in the default
branch, then the artifact must be closed if all requirements are met.

When this push occurs in another branch, then the artifact must not be
closed.

Merge Request approval to check if some artifacts must be closed after a
merge in the default branch will be handled in a dedicated feature commit.

Change-Id: I56d4d38be89dd0f17a46be64a9b382dd6837b6b5