refactoring: interface VerifyIsProgramIncrementTracker
part of story #21368 create mirrored iteration

No functional change expected

Code was building a full Tracker object when in fact all we needed to
check was that a tracker is (or is not) a program increment tracker.
Some of the checks did not properly catch exceptions too, so they would
bubble and result in errors.

Extracting this will simplify code when we build a
SourceTrackerCollection with a named constructor pattern

Change-Id: I5c5e4423a07b422820f902d44e2f213892cb1ef6