[feature] Extract refs in branch name
This is part of story #17181 reference tuleap artifacts in gitlab branches

If a pattern TULEAP-xxxx is found in the branch name, then a
cross-reference is saved. This is handled in the Post Push webhook.

The same branch referencing the same artfiact is not saved multiple
times.

The cross-reference is not yet displayed in the artifact view.

Change-Id: I4c8165301a4e39256bffd21c7238a59c4f5469f5