refactor: move time_period in store
Until now, the time period displayed was based on displayed tasks (from
the first start date to the last end date).

However, with the introduction of iterations, time period calculation
will change to start just before the iteration containing the first task,
and to end after the iteration containing the last task.

Since both tasks and iterations are managed by the store, it make sense
to move the build of the TimePeriod to the store. It has the advantage
to decrease a little bit the size of the main component GanttBoard.

Part of story #21379: Display iterations on top of roadmap

Change-Id: Ied88bfada539b7bfa18582d5e6cb6234a6c6cab3