feat: Display iterations in Roadmap widget
If a roadmap has defined one or two iteration trackers (you have to hack
the db for now), then corresponding iterations are displayed.

If iteration have end date < start date, then it won't be displayed.

If iteration is too short, then text won't be visible and corresponding
artifact won't be accessible. Since it is not the nominal case
(iterations usually span on a minimum 2 weeks window), we ignore this
case.

Part of story #21379: Display iterations on top of roadmap

Change-Id: I856a014507d47274ef24b59e52871256ddcec771