Give ProgramIdentifier to PI builder instead of Project
part of story #21368 create mirrored iteration

No functional change expected.

Replace the generic Project by ProgramIdentifier. ProgramIdentifier is
guaranteed to be a Program project, and is more explicit.
The "UserCannotSee" exception has been removed because it was never
distinguished from "NotFound" and from a user's perspective, it is the
same thing (if you can't see a project, it's the same as if it does not
exist).

Change-Id: I78940ef48e36d78660e693416d538d22dcb9086f