Git branch selector displays the default branch
Since it is not yet possible to change the default branch it only
display the badge on branches named "master" (it is the current git
default). You can still play with it by executing in a repo as the
gitolite user `git symbolic-ref HEAD refs/heads/<branch_name>`.

Part of request #13356: Allow to change the default branch

Change-Id: Id5ad1cf1d97c95f28ee46d022253263136c14c88