Refacto - Retrieve team trackers collection out of TimeboxCreatorChecker
This is part of story #21368 create mirrored iteration

No functional changes.

This patch move the recovery of Teams tracker in
ProgramIncrementCreatorChecker because TimeboxCreatorChecker will be
used in next patches by IterationCreatorChecker to check if an iteration
will can be created.
But since now, TimeboxCreatorChecker will retrieve only teams tracker of
Program Increment (not iteration).

How to test:
 - If you have right, you can still create a PI
 - If something is wrong (bad semantic, bad user right, ...) logs are
   logged

Note to reviewer:
 - I add a logger in ProgramIncrementCreatorChecker to know which
   milestone type we deal with
 - TimeboxCreatorCheckerTest use PHPUnit instead of Mockery

Change-Id: Id75306dc814a7393f44f8c74769f7727a863fe21