[feature] clean all data (information + cross-references) at integration deletion
This is part of story #17181 reference tuleap artifacts in gitlab branches

When removing an integration, only information where previously removed,
keeping useless cross-reference in database. They are now also removed.

When removing an integration of a GitLab repository integrated in
another projects, only the information and the cross-references of this
specific integration are removed.

Change-Id: I187d7f7f45bf96ed174fd203f69748c2423055ec