[feature] display last push date in GitLab branch cross-reference
This is part of story #17181 reference tuleap artifacts in gitlab branches

The last push date information is now saved for a referenced branch.
If this information is store for a reference, then it's displayed in the
cross-reference in the artifact.

Like the SHA-1 information, the branch's last push date is updated on
an already referenced branch.

Change-Id: If5905823080486d43414cf0187c4570f3bfe2905