feat: Don't display useless iterations
The important matter of the gantt visualisation are the tasks.
Iterations are just here to give context about those tasks. Therefore we
don't need to display iterations that are out of scope.

For example if you have one task displayed in your gantt with the
following dates: { start: 2020-03-10, end: 2020-04-10 }
and you have defined an iteration tracker for this widget (remember that
you have to hack the database for now) with the following iterations:
* Sprint A { start: 2020-01-01, end: 2020-01-31 }
* Sprint B { start: 2020-02-01, end: 2020-02-28 }
* Sprint C { start: 2020-03-01, end: 2020-03-31 }
* Sprint D { start: 2020-04-01, end: 2020-04-30 }
* Sprint E { start: 2020-05-01, end: 2020-05-31 }
* Sprint F { start: 2020-05-01, end: 2020-06-30 }

Then only Sprint C and Sprint D should be displayed.

The display of iterations should be adjusted to the display of subtasks.
If the task as a subtask with the following dates :
{ start: 2020-04-10, end: 2020-05-10 }
Then when the task is expanded, the Sprint E should appear.

Part of story #21379: Display iterations on top of roadmap

Change-Id: I8c33413c241d2da53ceaefd9855c136f62dbc93d