request #21772: Baseline plugin hardcodes admin accesses at plugin installation without the possibility to change them
The plugin assigns some roles during its installation with the current
instance state and give no possibility to change that or even to see
which groups can do operations with the plugin.

Assigning some roles by default is already dangerous in itself because
there is no way to know if it desirable to give access to a specific
group of users in the context of a project (it basically fails open at
some fixed point in time and hope for the best). It is even worth to do it
when administrators have no control over it.

This contribution drops all existing roles: only project administrators
or site administrators will be able to read or create baselines. If it
is necessary to give accesses to other users it will be necessary to
develop the missing permission administration feature (starting with the
possibility to see the service permissions in the project
administrations).

Change-Id: I0d3d015e75807c9810082793379e410546a1ff97