Ease SemanticTimeframeBuilder instanciation
Timeframes are used in various places in Tuleap, and each time we need
to inject a new dependency in SemanticTimeframeBuilder, we need to
update 24 locations in several plugins.

It became an actual pain in my a**: it's been two or three times
this release I had to update those files. Moreover it clutters
significantly the reviews.

part of: story #21362 Base artifact timeframe on a linked artifact

No functional change expected

Change-Id: I5834ae23ae056cf8fe06366c0f529bd0e2c51c4c