Block deletion of the default branch of a Git repository
Note this is already done by Git by default via the setting `receive.denyDeleteCurrent`.
Nonetheless, even if it is protected by default it would not be the
first time administrators play with settings they do not fully
comprehend so having a complementary safety net is not a bad thing.

To test:
 * you need to create the symlink to the new Gitolite VREF (checkout the
   specfile change to find out the source and target)
 * force a deployment of the Gitolite configuration (change a
   permissions in a repo somewhere, regenerate the conf of a project in
   site administration...)
 * disable the built-in Git protection in a repo (via `git config receive.denyDeleteCurrent ignore`)
   or globally

Part of request #13356: Allow to change the default branch

Change-Id: I724db34ee76d6cc835376bd36c0219fc50381544