Set default branch of Git repositories to main
This is mainly a technical change since Tuleap, as of today,
only creates empty repositories so when users clone it for the first
time their local config of init.defaultBranch has priority.

However this change makes us future proof:
 * we now have the possibility to explicitly choose the default branch
   name which can be interesting if we want to init repositories with
   some content
 * it makes sure we will not be surprised by future changes in Git
 * it allows us to remove the usage of the term "master" in some of the
   tests

To test you can create a new repository and check the result of
`git symbolic-ref HEAD` in the newly created reprository.

Part of request #13356: Allow to change the default branch

Change-Id: Ia97299aa3c0e195fda7527797a24c622893870a6