Reduce duplication between TCE and TEE docker start
Most of the run is shared so better extract it. The
main difference is that TCE install and activate
all plugins by default while TEE do not. This is because
we already have use case of TEE where not all plugins
should be activated by default (project_ownership...).

Part of request #21334 docker: replace enalean/tuleap-aio by tuleap/tuleap-community-edition

Change-Id: Ib6a7cebce94847c87fadc12c8fa040f7e0aca285