Set the default branch to 'main' when creating a new repository
The first attempt done in git #tuleap/stable/dffffec067b171d2dbfd80c51c2f80da98ffb8e2
resulted in broken Git repositories with the default global Git
settings. The issue was caused by the fact that while repositories are
created by the gitolite user, the change of the default branch was set
done as the codendiadm user which resulted in a file owned and only
readable by it.

To avoid this issue, the initial default branch modification when the
repo is created is now done by the gitolite user to avoid issues.

To test, do not forget to deploy the new sudoers config file.

It is possible to fix the repositories that were broken by the initial
implementation with the following command:

```
 # find /var/lib/gitolite/repositories/ -user codendiadm -name 'HEAD' -type f -exec sh -c 'chown gitolite:gitolite {} && chmod 0660 {}' \;
```

Part of request #13356: Allow to change the default branch

Change-Id: I3de44e4445a2df39b58488a5706492efa81222ee