vue: 2.6.13 -> 2.6.14
Changelog: https://github.com/vuejs/vue/releases/tag/v2.6.14

This is needed because due to the way our utility initVueGettext is
designed, we end relying on the signature of the Vue instance present
in @tuleap/core. The signature of the VueConstructor are not compatible
between Vue 2.6.13 and Vue 2.6.14 [0] which means TypeScript is not
happy if you currently try to bootstrap a new Vue app.

Part of story #21336: create a development branch on GitLab out of an artifact

[0] https://github.com/vuejs/vue/commit/c52427b0d2c1d203deea6eb69f2b4b181d56022c

Change-Id: I8ca7dcb55ed0ed8d8e56dddad7aba60c000fd6e6