Prevent tracker admins to reset the semantic if there are trackers relying on it
part of: story #21362 Base artifact timeframe on a linked artifact

To bring more guidance to the user, I now display the current
configuration in the admin UI of the semantic timeframe (inspired from
the velocity semantic).

How to test:

In the following test, I consider that the semantic timeframe of my User
Story tracker (US) is implied from the Sprint tracker.

- Go to the admin of the timeframe semantic of the Sprint tracker
--> You cannot reset the semantic
--> Message on button hover: "you cannot because blablabla"

- Remove required attributes on all fields select boxes
- Remove the disabled attribute on the reset button
- Reset the semantic
--> You cannot

NB: We should also probably block the reset backend-side when the
tracker has graphs depending on the configuration.

I'll do it in another contribution.

Change-Id: I9258e668f1569d378f29180ba9addcdd1edcd5c3