Add cog icon for FlamingParrot Project Admin
Part of story #21791 see Programs & Teams in sidebar

How to test:
- In FlamingParrot (go to Tracker reports), when you are project
administrator, there is a "cog" icon to the right of the project's
public name in the sidebar
- When you have a long project name, the name still wraps on two lines.
- The "Admin" service is still present and will be removed near the end
of the story because we must first display the cog icon in BurningParrot
- Other icons (project privacy, project banner button, project flags,
etc). will be changed in later commits to match the mockups.

Change-Id: Iab584aae86a6e763aa845020ec75f7beec178c14