Convert semantic-timeframe-option-selector to TS
part of: story #21362 Base artifact timeframe on a linked artifact

This refactoring is needed because I need to add additional behaviors to
the front-end of the semantic timeframe admin UI. Since all new
developments have to be written in TypeScript, this is the moment to
convert it.

Moreover, I took the opportunity to add the missing unit tests.

No functionnal change expected.

Change-Id: I7dfb65332b05f00261096dab3b2e0099be6746a2