feat: select more than one tracker for Roadmap
The roadmap widget can now be configured to track the progress of many
trackers at once. You can for example mix R&D and Marketing activities
at the same time with common context.

Hoping that the review will not be too big, this contribution is
complete meaning that feature is available:
* At widget creation
* At widget edition
* Via XML import
* Via project template duplication

Closes story #21807: select more than one tracker for Roadmap

Change-Id: I6cd594a9afa3870583c1109c26e4195a03efa5a7