Display Teams of Program in Sidebar (FlamingParrot)
part of story #21791 see Programs & Teams in sidebar

Entrypoint: FlamingParrot / BurningParrot Theme. They dispatch
CollectLinkedProjects event which is then handled by
CollectLinkedProjectsHandler in Program Management

How to test:
- Given you have configured a program project with some teams, when you
browse the Program project (only in FlamingParrot), you will see the
list of Teams in the expanded sidebar
- when the sidebar is collapsed, only the icon should be shown for now.
In later commits, a popover will show more information.
- when your current user cannot see a Team project (for example it's
Private and the user is not member), the Team is not listed.
- when your current user cannot see any Team project, the "linked
projects block" should not appear.
- when the project is not a program (it's a team, or anything else not
linked to program management), the "block" should not appear.

Notes to reviewer:
- In CollectLinkedProjects, there is a "parent" projects collection. It
is never filled for now but is added to give clarity to the
architecture. In later commits, the Handler will add to this collection
when current project is a Team. At that time, the plan is to introduce a
"ProgramsSearcher" and to build a collection with it.
- I know I introduced a lot of interfaces, but they allow us to write
almost mock-free tests. Without "RetrieveProject", I would have no
choice but to mock ProjectManager. Same with "CheckProjectAccess".
- The mockups don't tell us if the Teams are links, I assumed they were
links to the respective project's dashboard.

Change-Id: I37d1e64359eea0975623e08a91e234c27cf8217f