Add a GET route on /git/:id/tree
This change is part of story #18320 native support of pull requests in tuleap branch source jenkins plugin

The goal of this route is to be able to retrieve the tree content of
a path on a Git repository.

An error should be thrown if reference or folder is not found or if
the provided path is not one of a directory.

Change-Id: I4bc96ccf7b60656474134cf2d597addf1c9dcd2d