[feature] Create the branch with the modal
This is part of story #21336: create a development branch on GitLab out of an artifact

If you fill the form, then the branch must be created on GitLab side.
The modal is not closed when the actions is done.

Errors are not handled yet. There is no visual feedback that shows that
actions is processing and finished. Those will be added in the next commit.

Change-Id: I5a3b8de0adbfb6a54e982b65befda199fd43e5ee