Stop using enalean/tuleap-reverseproxy image
This image is only used in the E2E tests of the distributed Tuleap
setup. Building the image when needed instead of pulling it does
not make a big difference anyway (most of the time is spent running
the tests anyway) and it is easier to maintain this way.

The Dockerfile has been copied as is with only a small change so we can
cache on our infra the base image.

Part of request #21809: Stop relying on Docker Hub autobuild feature

Change-Id: I55b8b22f404debafbf16b6952cf6f75d9044b78e