Semantic timeframe admin as a Vue app
part of: story #21362 Base artifact timeframe on a linked artifact

Motivation:

The complexity of the UX will be increased in the next commit, making it
hard to continue with only Typescript and mustache. A vue.js app will be
a lot easier to maintain and to extend.

No functional change expected

Change-Id: If6c7658502309dbe85dae9e1c8d86602e3003650