fix: progress for roadmap tasks without artifact link value.
Create a roadmap task with CSV import. There is a high chance that it
won't have any artifact link value for the artifact link field.

If tracker is configured to have a progress "Count based", then the
roadmap widget should display a progress (either 0% or 100% depending on
the artifact status), just like if the artifact would have an artifact
link value without any links.

Closes request #21813: Roadmap Task without artifact link value should have a progress displayed

Change-Id: I7a579720af8ffab7954691e189ef58cece3e9b58