Split SearchTeamsOfProgram interface
part of story #21791 see Programs & Teams in sidebar

No functional change expected.

ProgramStore did a lot of things, which makes it harder to stub.
Besides, it returned an array of array, which does not make it easy to
use. Returning an array of int is easier for callers.

It will be useful for us to build a list of team projects in the
sidebar.

Change-Id: I92a7f0555ab9c61f3040cca9b9809cd89a70294d