Split VerifyIsProgram interface
part of story #21791 see Programs & Teams in sidebar

No functional change expected.

ProgramStore did a lot of things, which makes it harder to stub. Using a
smaller interface allows us to replace a lot of mocks with a simple stub
class.

It will be useful for us to build a list of team projects in the
sidebar.

Change-Id: I17041c8e81c93eb191a69e016a4531def6edd811