Enforce branch name when creating a branch on GitLab from an artifact
There is no good reason for the API on the Tuleap side to let users
create branch on a GitLab with arbitrary names. It opens the possibility
of abuses. If users want to create branch with arbitrary name, they
should call the GitLab API directly.

Also the frontend unit tests are now played by the CI pipeline.

Part of story #21336: create a development branch on GitLab out of an artifact

Change-Id: I69ab8a37f4c97c819750d3919c234e48df10ab59