Information about branches of the integrated GitLab repositories can be requested
The Tuleap REST API now exposes a GET /gitlab_repositories/:id/branches
to retrieve information about branches of the integrated GitLab
repositories. For now, only the default branch is returned since it is
the only information Tuleap needs.

Part of story #21336: create a development branch on GitLab out of an artifact

Change-Id: I8b1c24c7408d69e5d6a4a79c13e0b219a1c3e06e