Extract concurrency limit pool from TestPlan
The behavior is extracted into a dedicated library as it something that
can be reused in multiple places and have the potential to be expanded
to fulfill other needs (e.g. different resolution algorithme).

We are going to have the same exact needs in the GitLab create branch
modal to retrieve all the default branches of GL repositories in a
efficient way.

Part of story #21336: create a development branch on GitLab out of an artifact

Change-Id: Id91c2b3addd6b56211ea1cf575fab1b28f8b789c