request #21826 : Quickly access Taskboard content
When navigating using the keyboard in Taskboard, it can be painful to reach the actual content of the Taskboard.
Even if using the "go to main content" button, there are still some elements between the top of the page's main content and the first swimlane or card.
In addition, the first element of a swimlane is not the first card but the toggle button which can be confusing.

Two new shortcuts are added to help solve this issue :
* When on a swimlane, pressing `enter` will focus the first card.
* Pressing `ctrl+up` will place focus on the first swimlane.

That way, it is possible to reach the content more quickly, but
secondary actions are still available to keyboard users.

This patch also adds the possibility, inside the KeyboardShortcuts lib, to prevent browser default behaviour or not when using a
shortcut.
In that case, `ctrl+up` will focus the first swimlane, but if focus is
already on first swimlane, it will not prevent the default browser
behaviour, which is to scroll back to the top of the page.

Change-Id: I2861eb8aa67f7dd021e413559ffd9ac75aa0b766