Prefill default branch of the GitLab repository in the "create branch" modal
We make an attempt to retrieve the default branch of the GitLab
repositories to prefill the source field.
If for some reasons that fail (e.g. the GitLab API is too slow) we just
skip over and do not prefill it as the branch creation itself might
still work.

Part of story #21336: create a development branch on GitLab out of an artifact

Change-Id: I067b2ea6a1e02d812050b4b2d9aff76ca33994b9