request #21851 : Focus is lost when closing modal
When navigating with the keyboard, if I open and close a modal, focus is lost.
This patch makes the focus go back to where it was before opening the
modal.

Note: For now, it does not work with Flaming Parrot modals, as they use
Bootstrap. It should be fixed in another patch.

Change-Id: I7ffe984a8d6c4987cd57c247e81dc667dabb0225