[feature] Add Roadmap widget at Jira project import
This fixes request #21836 Add Roadmap widget to projects imported from Jira

If the Epic issueType is found and if the Tuleap timeframe semantic can
be created for it (this means having both start date and due date
fields), then a new project dashboard named "Roadmap" is created.

A new dependency to the roadmap plugin has been added.

Change-Id: I23f0e0a92367c8f70c766e510a8258eaae232474